move queuing logic into a new file in main package
